Make repititive code into hooks and misc changes
