+++ src/api/index.js
@@ -4,9 +4,9 @@ const getShareURL = (ticket_no, type) => {
             return `${process.env.REACT_APP_SHARE_ENDPOINT}/team/ticket/${ticket_no}`;
         case "awareness":
             return `${process.env.REACT_APP_SHARE_ENDPOINT}/awareness-session/ticket/${ticket_no}`;
+        default:
+            return ""
     }
-
-    return "";
 }
 
 export { getShareURL };
